functions/python/image_analyzer/requirements.txt
# Core requirements
functions-framework>=3.0.0,<4.0.0
firebase-admin>=6.0.0,<7.0.0
openai>=1.0.0,<2.0.0
google-cloud-storage>=2.0.0,<3.0.0

# Image processing
opencv-python-headless>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
numpy>=1.0.0,<2.0.0

# Utils
python-dotenv>=1.0.0,<2.0.0 